# cmd/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "飞宏IM接口",
        "title": "飞宏IM接口",
        "contact": {
            "name": "飞宏IM技术支持",
            "url": "https://github.com/liony823/chat",
            "email": "ilovecoding@foxmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:10002",
    "basePath": "/",
    "paths": {
        "/group/cancel_mute_group": {
            "post": {
                "description": "取消禁言群组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "取消禁言群组",
                "operationId": "cancelMuteGroup",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.CancelMuteGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.CancelMuteGroupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/dismiss_group": {
            "post": {
                "description": "解散群组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "解散群组",
                "operationId": "dismissGroup",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.DismissGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.DismissGroupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/get_group_member_list": {
            "post": {
                "description": "获取群组成员列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "获取群组成员列表",
                "operationId": "getGroupMemberList",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.GetGroupMemberListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.GetGroupMemberListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/get_groups": {
            "post": {
                "description": "获取群组列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "获取群组列表",
                "operationId": "getGroups",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.GetGroupsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.GetGroupsResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/kick_group": {
            "post": {
                "description": "踢出群组成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "踢出群组成员",
                "operationId": "kickGroupMember",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.KickGroupMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.KickGroupMemberResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/mute_group": {
            "post": {
                "description": "禁言群组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "禁言群组",
                "operationId": "muteGroup",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.MuteGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.MuteGroupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/statistics/group/create": {
            "post": {
                "description": "获取群组统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "获取群组统计",
                "operationId": "groupCreateCount",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.GroupCreateCountReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/group.GroupCreateCountResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
        "group.CMSGroup": {
            "type": "object",
            "properties": {
                "groupInfo": {
                    "$ref": "#/definitions/sdkws.GroupInfo"
                },
                "groupOwnerUserID": {
                    "type": "string"
                },
                "groupOwnerUserName": {
                    "type": "string"
                }
            }
        },
        "group.CancelMuteGroupReq": {
            "type": "object",
            "properties": {
                "groupID": {
                    "type": "string"
                }
            }
        },
        "group.CancelMuteGroupResp": {
            "type": "object"
        },
        "group.DismissGroupReq": {
            "type": "object",
            "properties": {
                "deleteMember": {
                    "type": "boolean"
                },
                "groupID": {
                    "type": "string"
                }
            }
        },
        "group.DismissGroupResp": {
            "type": "object"
        },
        "group.GetGroupMemberListReq": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "integer"
                },
                "groupID": {
                    "type": "string"
                },
                "keyword": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/sdkws.RequestPagination"
                }
            }
        },
        "group.GetGroupMemberListResp": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sdkws.GroupMemberFullInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "group.GetGroupsReq": {
            "type": "object",
            "properties": {
                "groupID": {
                    "type": "string"
                },
                "groupName": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/sdkws.RequestPagination"
                }
            }
        },
        "group.GetGroupsResp": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/group.CMSGroup"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "group.GroupCreateCountReq": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "group.GroupCreateCountResp": {
            "type": "object",
            "properties": {
                "before": {
                    "type": "integer"
                },
                "count": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "group.KickGroupMemberReq": {
            "type": "object",
            "properties": {
                "groupID": {
                    "type": "string"
                },
                "kickedUserIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "group.KickGroupMemberResp": {
            "type": "object"
        },
        "group.MuteGroupReq": {
            "type": "object",
            "properties": {
                "groupID": {
                    "type": "string"
                }
            }
        },
        "group.MuteGroupResp": {
            "type": "object"
        },
...
        "sdkws.GroupInfo": {
            "type": "object",
            "properties": {
                "applyMemberFriend": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "creatorUserID": {
                    "type": "string"
                },
                "ex": {
                    "type": "string"
                },
                "faceURL": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "groupName": {
                    "type": "string"
                },
                "groupType": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "lookMemberInfo": {
                    "type": "integer"
                },
                "memberCount": {
                    "type": "integer"
                },
                "needVerification": {
                    "type": "integer"
                },
                "notification": {
                    "type": "string"
                },
                "notificationUpdateTime": {
                    "type": "integer"
                },
                "notificationUserID": {
                    "type": "string"
                },
                "ownerUserID": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "sdkws.GroupMemberFullInfo": {
            "type": "object",
            "properties": {
                "appMangerLevel": {
                    "description": "if \u003e0",
                    "type": "integer"
                },
                "ex": {
                    "type": "string"
                },
                "faceURL": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "inviterUserID": {
                    "type": "string"
                },
                "joinSource": {
                    "type": "integer"
                },
                "joinTime": {
                    "type": "integer"
                },
                "muteEndTime": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "operatorUserID": {
                    "type": "string"
                },
                "roleLevel": {
                    "type": "integer"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "sdkws.RequestPagination": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "showNumber": {
                    "type": "integer"
                }
            }
        },
...
    }
}